[BE] refactor: Redis 공통 설정 Configuration을 core 패키지로 이동 #817
