+++ backend/app/build.gradle
@@ -27,10 +27,6 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-validation'
     implementation 'org.springframework.boot:spring-boot-starter-log4j2'
 
-    // Redis & Redisson
-    implementation 'org.springframework.boot:spring-boot-starter-data-redis'
-    implementation 'org.redisson:redisson-spring-boot-starter:3.52.0'
-
     // Resilience4j (Circuit Breaker)
     implementation 'io.github.resilience4j:resilience4j-spring-boot3:2.3.0'
